Added blocks as horizontal row image
Using a grey block and changing the color with setColor won't work because it multiplies the image by the color value, so the tetris blocks look really dark. With this method, we can set the tetris block to be a certain part of the image and get a nice color.