Inheritance and Polymorphism
This are 2 classes. Game class is a base class that will do basic game functions like keep track of gamestate, levels, horizontal movement. also it will have a virtual function for polymorphism that checks if the player won the game.
tetris class conducts gameplay and has all tetris specific functionalites.
this is an initial skeleton